% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_mortality}
\alias{analyse_mortality}
\title{Analyse Detections and Recaptures}
\usage{
analyse_mortality(data, niters = 10^5)
}
\arguments{
\item{data}{A \code{analysis_data} object of the detection and recapture data to analyse.}

\item{niters}{An integer of the minimum number of MCMC iterations to
perform.}
}
\value{
A jags_analysis object.
}
\description{
Analyses detection and recapture data using a mortality model.
}
\details{
To view the full model description
in the JAGS dialect of the BUGS language use \code{\link{mortality_model_code}}.

The data must be a \code{analysis_data} object as generated by the
\code{\link{lexr::make_analysis_data}} function.
}

