% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_survival}
\alias{analyse_survival}
\title{Analyse Detections and Recaptures}
\usage{
analyse_survival(data, niters = 10^5, mode = "current")
}
\arguments{
\item{data}{A \code{analysis_data} object of the detection and recapture data to analyse.}

\item{niters}{An integer of the minimum number of MCMC iterations to
perform.}

\item{A}{character element indicating the mode for the analysis.}
}
\value{
A jags_analysis object.
}
\description{
Analyses detection and recapture data using a Bayesian individual multistate
state-space formulation of the Cormack-Jolly-Seber (CJS) survival model
}
\details{
To view the full model description
in the JAGS dialect of the BUGS language use \code{\link{survival_model_code}}.

The data must be an \code{analysis_data} object as generated by the
\code{\link{make_analysis_data}} function.
}

