% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_tagloss}
\alias{analyse_tagloss}
\title{Analyse Tag Loss}
\usage{
analyse_tagloss(table, niters = 10^3, mode = "current")
}
\arguments{
\item{table}{A table of the number of recaptures with $100 and $10 tags.}

\item{niters}{An integer of the minimum number of MCMC iterations to
perform.}

\item{mode}{A character element indicating the mode for the analysis.}
}
\value{
A jags_analysis object.
}
\description{
Analyses tag loss data using a simple Bayesian zero-truncated binomial tag-loss model.
}
\details{
To view the full model description
in the JAGS dialect of the BUGS language use \code{\link{tagloss_model_code}}.
}

