% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_analysis_length}
\alias{plot_analysis_length}
\title{Plot Analysis Length}
\usage{
plot_analysis_length(data, years = 2008:2013)
}
\arguments{
\item{data}{The data to plot.}

\item{years}{A numeric vector of the number of years.}
}
\value{
A ggplot2 object.
}
\description{
Plots analysis length.
}

