% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_mortality}
\alias{plot_mortality}
\title{Plot Mortality}
\usage{
plot_mortality(data1, data2, x, xlab = x, ylab = "Mortality")
}
\arguments{
\item{data1}{The first data to plot.}

\item{data2}{The second data to plot.}

\item{x}{A string of the column to plot on the x-axis.}

\item{xlab}{A string of the x-axis name.}

\item{ylab}{A string of the y-axis name.}
}
\value{
A ggplot2 object.
}
\description{
Plots the mortality of x
}

