% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pointrange}
\alias{plot_pointrange}
\title{Plot Point Range}
\usage{
plot_pointrange(data, x)
}
\arguments{
\item{data}{The data frame to plot.}

\item{x}{A string specifying the column to plot on the x-axis.}
}
\value{
A ggplot object.
}
\description{
Plot Point Range
}

