% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_probs}
\alias{plot_probs}
\title{Plot Probabilities}
\usage{
plot_probs(data1, data2)
}
\arguments{
\item{data}{The data to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plots the key probabilities.
}

