% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_section}
\alias{plot_section}
\title{Plot Section}
\usage{
plot_section(data)
}
\arguments{
\item{data}{The \code{lex_data} object to plot.}
}
\value{
A ggplot2 object.
}
\description{
Maps and labels the sections for the klexdatr  \code{lex_data} object.
}

