% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_spawning}
\alias{plot_spawning}
\title{Plot Spawning}
\usage{
plot_spawning(data1, data2)
}
\arguments{
\item{data}{The data to plot.}
}
\value{
A ggplot2 object.
}
\description{
Plots the probability of spawning by length.
}

