% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot Time Series}
\usage{
plot_timeseries(data, y, x = "Year", ylab = y, xlab = x, from = 2008,
  to = 2013, color = "black")
}
\arguments{
\item{data}{The first data to plot.}

\item{y}{A string of the column to plot on the y-axis.}

\item{x}{A string of the column to plot on the x-axis.}

\item{ylab}{A string of the y-axis name.}

\item{xlab}{A string of the x-axis name.}

\item{from}{A count of the start year.}

\item{to}{A count of the end year.}

\item{color}{A string of the color.}
}
\value{
A ggplot2 object.
}
\description{
Plots a time series
}

