% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_probs}
\alias{predict_probs}
\title{Predict Probabilities}
\usage{
predict_probs(analysis)
}
\arguments{
\item{analysis}{The jags_analysis object to predict the probabilities for.}
}
\value{
A data frame of probabilities
}
\description{
Predicts the key probabilities.
}

