% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{season}
\alias{season}
\title{Season}
\usage{
season(x)
}
\arguments{
\item{x}{A date-time object.}
}
\value{
An ordered factor with levels Winter, Spring, Summer and Autumn
}
\description{
Season
}
\examples{
season(1:12)
season(Sys.Date())
}

