% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawning.R
\name{spawning_rb}
\alias{spawning_rb}
\title{Rainbow Trout Spawning}
\usage{
spawning_rb(detection, period)
}
\arguments{
\item{detection}{A data.frame of the detection data for the capture.}

\item{period}{A data.frame of the periods.}
}
\description{
Identifies rainbow trout spawning events.
}

