% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{subtract600divide100}
\alias{subtract600divide100}
\title{Subtract 600 Divide 100}
\usage{
subtract600divide100(x)
}
\arguments{
\item{x}{The numeric vector to transform.}
}
\description{
Subtract 600 Divide 100
}
\examples{
subtract600divide100(c(400,500,600,750))
}

