% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{subtract750}
\alias{subtract750}
\title{Subtract 600 Divide 100}
\usage{
subtract750(x)
}
\arguments{
\item{x}{The numeric vector to transform.}
}
\description{
Subtract 600 Divide 100
}
\examples{
subtract750(c(400,500,600,750))
}

