% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{subtract750divide100}
\alias{subtract750divide100}
\title{Subtract 750 Divide 100}
\usage{
subtract750divide100(x)
}
\arguments{
\item{x}{The numeric vector to transform.}
}
\description{
Subtract 750 Divide 100
}
\examples{
subtract750divide100(c(400,500,600,750))
}

