% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise Results}
\usage{
summarise_results(lex, detect, bull_trout, rainbow_trout)
}
\arguments{
\item{lex}{The hourly raw data}

\item{detect}{The daily detection data}

\item{bull_trout}{The bull trout analysis object}

\item{rainbow_trout}{The rainbow trout analysis object}
}
\value{
A named list of key results.
}
\description{
Summarises key results as a named list.
}

