% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{survival_model_code}
\alias{survival_model_code}
\title{Survival Model Code}
\usage{
survival_model_code(model = "final", sd = 3, comments = TRUE)
}
\arguments{
\item{model}{A string specifying the model type ("base", "full" and "final")}

\item{sd}{A number specifying the standard deviation for the prior distributions.}

\item{comments}{A flag indicating whether to include comments.}
}
\value{
A string of the JAGS model code.
}
\description{
Returns a string of the JAGS code
defining the survival model.
}
\examples{
cat(survival_model_code())
}

