% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{tagloss_model_code}
\alias{tagloss_model_code}
\title{Tag Loss Model Code}
\usage{
tagloss_model_code(comments = TRUE)
}
\arguments{
\item{comments}{A flag indicating whether to include comments.}
}
\value{
A string of the JAGS model code.
}
\description{
Returns a string of the JAGS code
defining the tag loss model.
}
\examples{
cat(tagloss_model_code())
}

